+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+- Change `CallableTree::Node::Internal#append` to return a new instance.
+  To keep the same behavior as the older version, use `CallableTree::Node::External#append!` that make destructive change.
+- Remove `CallableTree::Node::Internal#<<`. Use `CallableTree::Node::External#append!` instead.
 - Change `CallableTree::Node::External#verbosify` to return a new instance.
   To keep the same behavior as the older version, use `CallableTree::Node::External#verbosify!` that make destructive change.